refactored from repository to factory app seems to still work
